Module Stdppx.Char

include module type of struct include Char end

val code : char -> int

Return the ASCII code of the argument.

val chr : int -> char

Return the character with the given ASCII code.

raises Invalid_argument
if the argument is outside the range 0--255.

val escaped : char -> string

Return a string representing the given character, with special characters escaped following the lexical conventions of OCaml. All characters outside the ASCII printable range (32..126) are escaped, as well as backslash, double-quote, and single-quote.

val lowercase_ascii : char -> char

Convert the given character to its equivalent lowercase character, using the US-ASCII character set.

val uppercase_ascii : char -> char

Convert the given character to its equivalent uppercase character, using the US-ASCII character set.

type t = char

An alias for the type of characters.

val seeded_hash : int -> t -> int

A seeded hash function for characters, with the same output value as Hashtbl.seeded_hash. This function allows this module to be passed as argument to the functor Hashtbl.MakeSeeded.

val hash : t -> int

An unseeded hash function for characters, with the same output value as Hashtbl.hash. This function allows this module to be passed as argument to the functor Hashtbl.Make.

include Comparisons with type t := char

val compare : char -> char -> int

val equal : char -> char -> bool

val (=) : char -> char -> bool

val (<) : char -> char -> bool

val (>) : char -> char -> bool

val (<>) : char -> char -> bool

val (<=) : char -> char -> bool

val (>=) : char -> char -> bool

val min : char -> char -> char

val max : char -> char -> char